toolchain+libc: Fixed libc related build errors
 - moved all headers files from the include folder
   to the root libc directory
 - Changed Binutils and GCC versions
 - Added time.h
 - Changed word ('unsigned short') to unsigned int
 - Added memset to string.h

+++ lib/libc/sys/types.h
@@ -5,7 +5,7 @@
 __BEGIN_DECLS
 
 typedef unsigned int dword;
-typedef unsigned short word;
+typedef unsigned int word;
 typedef unsigned char byte;
 
 typedef signed int signed_dword;
